convert module to adapter pattern
